Rename web components folders.
