+++ src/js/web-components/progress-ring/ProgressRing.js
